Replace nivel.id type from SMALLINT to INT

+++ pkg/denominacion-5R9BK7DS01O7A23G/upgrade.sql
@@ -1,7 +1,7 @@
 CREATE OR REPLACE TABLE denominacion (
     id TINYINT AUTO_INCREMENT,
     code TINYINT NOT NULL,  -- id at simo
-    nivel_id SMALLINT UNSIGNED,
+    nivel_id INT UNSIGNED,
     nombre VARCHAR(100) NOT NULL,
 
     PRIMARY KEY(id),
